+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Changed from transients to WP Object Cache in `LangInterface`'s `get_post_language` and `get_term_language`. The amount of insert/delete transient queries caused significant slowdown in pages with a lot of elements.
+
 ## [0.5.8] - 2025-01-13
 
 ### Added